refactor: create mocks module with mocklshw
